requirements.txt
aiofiles==24.1.0
httpx==0.28.1
jinja2==3.1.6
pyyaml==6.0.2
markdown==3.8
